+++ modules/apps/rate-limiting/types/query.pb.go
@@ -6,6 +6,7 @@ package types
 import (
 	context "context"
 	fmt "fmt"
+	query "github.com/cosmos/cosmos-sdk/types/query"
 	_ "github.com/cosmos/gogoproto/gogoproto"
 	grpc1 "github.com/cosmos/gogoproto/grpc"
 	proto "github.com/cosmos/gogoproto/proto"
@@ -31,6 +32,7 @@ const _ = proto.GoGoProtoPackageIsVersion3 // please upgrade the proto package
 
 // Queries all rate limits
 type QueryAllRateLimitsRequest struct {
+	Pagination *query.PageRequest `protobuf:"bytes,1,opt,name=pagination,proto3" json:"pagination,omitempty"`
 }
 
 func (m *QueryAllRateLimitsRequest) Reset()         { *m = QueryAllRateLimitsRequest{} }
@@ -66,9 +68,17 @@ func (m *QueryAllRateLimitsRequest) XXX_DiscardUnknown() {
 
 var xxx_messageInfo_QueryAllRateLimitsRequest proto.InternalMessageInfo
 
+func (m *QueryAllRateLimitsRequest) GetPagination() *query.PageRequest {
+	if m != nil {
+		return m.Pagination
+	}
+	return nil
+}
+
 // QueryAllRateLimitsResponse returns all the rate limits stored on the chain.
 type QueryAllRateLimitsResponse struct {
-	RateLimits []RateLimit `protobuf:"bytes,1,rep,name=rate_limits,json=rateLimits,proto3" json:"rate_limits"`
+	RateLimits []RateLimit         `protobuf:"bytes,1,rep,name=rate_limits,json=rateLimits,proto3" json:"rate_limits"`
+	Pagination *query.PageResponse `protobuf:"bytes,2,opt,name=pagination,proto3" json:"pagination,omitempty"`
 }
 
 func (m *QueryAllRateLimitsResponse) Reset()         { *m = QueryAllRateLimitsResponse{} }
@@ -107,6 +117,13 @@ var xxx_messageInfo_QueryAllRateLimitsResponse proto.InternalMessageInfo
 func (m *QueryAllRateLimitsResponse) GetRateLimits() []RateLimit {
 	if m != nil {
 		return m.RateLimits
+	}
+	return nil
+}
+
+func (m *QueryAllRateLimitsResponse) GetPagination() *query.PageResponse {
+	if m != nil {
+		return m.Pagination
 	}
 	return nil
 }
@@ -211,7 +228,8 @@ func (m *QueryRateLimitResponse) GetRateLimit() *RateLimit {
 
 // Queries all the rate limits for a given chain
 type QueryRateLimitsByChainIDRequest struct {
-	ChainId string `protobuf:"bytes,1,opt,name=chain_id,json=chainId,proto3" json:"chain_id,omitempty"`
+	ChainId    string             `protobuf:"bytes,1,opt,name=chain_id,json=chainId,proto3" json:"chain_id,omitempty"`
+	Pagination *query.PageRequest `protobuf:"bytes,2,opt,name=pagination,proto3" json:"pagination,omitempty"`
 }
 
 func (m *QueryRateLimitsByChainIDRequest) Reset()         { *m = QueryRateLimitsByChainIDRequest{} }
@@ -254,9 +272,17 @@ func (m *QueryRateLimitsByChainIDRequest) GetChainId() string {
 	return ""
 }
 
+func (m *QueryRateLimitsByChainIDRequest) GetPagination() *query.PageRequest {
+	if m != nil {
+		return m.Pagination
+	}
+	return nil
+}
+
 // QueryRateLimitsByChainIDResponse returns all rate-limits by a chain.
 type QueryRateLimitsByChainIDResponse struct {
-	RateLimits []RateLimit `protobuf:"bytes,1,rep,name=rate_limits,json=rateLimits,proto3" json:"rate_limits"`
+	RateLimits []RateLimit         `protobuf:"bytes,1,rep,name=rate_limits,json=rateLimits,proto3" json:"rate_limits"`
+	Pagination *query.PageResponse `protobuf:"bytes,2,opt,name=pagination,proto3" json:"pagination,omitempty"`
 }
 
 func (m *QueryRateLimitsByChainIDResponse) Reset()         { *m = QueryRateLimitsByChainIDResponse{} }
@@ -299,9 +325,17 @@ func (m *QueryRateLimitsByChainIDResponse) GetRateLimits() []RateLimit {
 	return nil
 }
 
+func (m *QueryRateLimitsByChainIDResponse) GetPagination() *query.PageResponse {
+	if m != nil {
+		return m.Pagination
+	}
+	return nil
+}
+
 // Queries all the rate limits for a given channel or client ID
 type QueryRateLimitsByChannelOrClientIDRequest struct {
-	ChannelOrClientId string `protobuf:"bytes,1,opt,name=channel_or_client_id,json=channelOrClientId,proto3" json:"channel_or_client_id,omitempty"`
+	ChannelOrClientId string             `protobuf:"bytes,1,opt,name=channel_or_client_id,json=channelOrClientId,proto3" json:"channel_or_client_id,omitempty"`
+	Pagination        *query.PageRequest `protobuf:"bytes,2,opt,name=pagination,proto3" json:"pagination,omitempty"`
 }
 
 func (m *QueryRateLimitsByChannelOrClientIDRequest) Reset() {
@@ -348,9 +382,17 @@ func (m *QueryRateLimitsByChannelOrClientIDRequest) GetChannelOrClientId() strin
 	return ""
 }
 
+func (m *QueryRateLimitsByChannelOrClientIDRequest) GetPagination() *query.PageRequest {
+	if m != nil {
+		return m.Pagination
+	}
+	return nil
+}
+
 // QueryRateLimitsByChannelOrClientIDResponse returns all rate-limits by a channel or client id.
 type QueryRateLimitsByChannelOrClientIDResponse struct {
-	RateLimits []RateLimit `protobuf:"bytes,1,rep,name=rate_limits,json=rateLimits,proto3" json:"rate_limits"`
+	RateLimits []RateLimit         `protobuf:"bytes,1,rep,name=rate_limits,json=rateLimits,proto3" json:"rate_limits"`
+	Pagination *query.PageResponse `protobuf:"bytes,2,opt,name=pagination,proto3" json:"pagination,omitempty"`
 }
 
 func (m *QueryRateLimitsByChannelOrClientIDResponse) Reset() {
@@ -397,8 +439,16 @@ func (m *QueryRateLimitsByChannelOrClientIDResponse) GetRateLimits() []RateLimit
 	return nil
 }
 
+func (m *QueryRateLimitsByChannelOrClientIDResponse) GetPagination() *query.PageResponse {
+	if m != nil {
+		return m.Pagination
+	}
+	return nil
+}
+
 // Queries all blacklisted denoms
 type QueryAllBlacklistedDenomsRequest struct {
+	Pagination *query.PageRequest `protobuf:"bytes,1,opt,name=pagination,proto3" json:"pagination,omitempty"`
 }
 
 func (m *QueryAllBlacklistedDenomsRequest) Reset()         { *m = QueryAllBlacklistedDenomsRequest{} }
@@ -434,9 +484,17 @@ func (m *QueryAllBlacklistedDenomsRequest) XXX_DiscardUnknown() {
 
 var xxx_messageInfo_QueryAllBlacklistedDenomsRequest proto.InternalMessageInfo
 
+func (m *QueryAllBlacklistedDenomsRequest) GetPagination() *query.PageRequest {
+	if m != nil {
+		return m.Pagination
+	}
+	return nil
+}
+
 // QueryAllBlacklistedDenomsResponse returns all the blacklisted denosm.
 type QueryAllBlacklistedDenomsResponse struct {
-	Denoms []string `protobuf:"bytes,1,rep,name=denoms,proto3" json:"denoms,omitempty"`
+	Denoms     []string            `protobuf:"bytes,1,rep,name=denoms,proto3" json:"denoms,omitempty"`
+	Pagination *query.PageResponse `protobuf:"bytes,2,opt,name=pagination,proto3" json:"pagination,omitempty"`
 }
 
 func (m *QueryAllBlacklistedDenomsResponse) Reset()         { *m = QueryAllBlacklistedDenomsResponse{} }
@@ -479,8 +537,16 @@ func (m *QueryAllBlacklistedDenomsResponse) GetDenoms() []string {
 	return nil
 }
 
+func (m *QueryAllBlacklistedDenomsResponse) GetPagination() *query.PageResponse {
+	if m != nil {
+		return m.Pagination
+	}
+	return nil
+}
+
 // Queries all whitelisted address pairs
 type QueryAllWhitelistedAddressesRequest struct {
+	Pagination *query.PageRequest `protobuf:"bytes,1,opt,name=pagination,proto3" json:"pagination,omitempty"`
 }
 
 func (m *QueryAllWhitelistedAddressesRequest) Reset()         { *m = QueryAllWhitelistedAddressesRequest{} }
@@ -516,9 +582,17 @@ func (m *QueryAllWhitelistedAddressesRequest) XXX_DiscardUnknown() {
 
 var xxx_messageInfo_QueryAllWhitelistedAddressesRequest proto.InternalMessageInfo
 
+func (m *QueryAllWhitelistedAddressesRequest) GetPagination() *query.PageRequest {
+	if m != nil {
+		return m.Pagination
+	}
+	return nil
+}
+
 // QueryAllWhitelistedAddressesResponse returns all whitelisted pairs.
 type QueryAllWhitelistedAddressesResponse struct {
 	AddressPairs []WhitelistedAddressPair `protobuf:"bytes,1,rep,name=address_pairs,json=addressPairs,proto3" json:"address_pairs"`
+	Pagination   *query.PageResponse      `protobuf:"bytes,2,opt,name=pagination,proto3" json:"pagination,omitempty"`
 }
 
 func (m *QueryAllWhitelistedAddressesResponse) Reset()         { *m = QueryAllWhitelistedAddressesResponse{} }
@@ -561,6 +635,13 @@ func (m *QueryAllWhitelistedAddressesResponse) GetAddressPairs() []WhitelistedAd
 	return nil
 }
 
+func (m *QueryAllWhitelistedAddressesResponse) GetPagination() *query.PageResponse {
+	if m != nil {
+		return m.Pagination
+	}
+	return nil
+}
+
 func init() {
 	proto.RegisterType((*QueryAllRateLimitsRequest)(nil), "ibc.applications.rate_limiting.v1.QueryAllRateLimitsRequest")
 	proto.RegisterType((*QueryAllRateLimitsResponse)(nil), "ibc.applications.rate_limiting.v1.QueryAllRateLimitsResponse")
@@ -581,54 +662,60 @@ func init() {
 }
 
 var fileDescriptor_f55a91bf266ae0f7 = []byte{
-	// 742 bytes of a gzipped FileDescriptorProto
-	0x1f, 0x8b, 0x08, 0x00, 0x00, 0x00, 0x00, 0x00, 0x02, 0xff, 0xb4, 0x56, 0x41, 0x4f, 0x13, 0x5b,
-	0x14, 0xee, 0xf0, 0x1e, 0xbc, 0xd7, 0xc3, 0x63, 0xf1, 0xae, 0x15, 0x61, 0xd4, 0x02, 0xa3, 0x44,
-	0x34, 0xb6, 0x93, 0x62, 0x8c, 0xa2, 0x10, 0x43, 0x4b, 0x50, 0x22, 0x22, 0xd6, 0x44, 0x12, 0x63,
-	0x6c, 0x6e, 0x67, 0x6e, 0xda, 0x1b, 0xa7, 0x73, 0x87, 0xb9, 0x53, 0x48, 0x63, 0x58, 0xe8, 0x9a,
-	0x85, 0x89, 0x3f, 0xc7, 0x3f, 0xc0, 0x92, 0xc4, 0x0d, 0x2b, 0x63, 0x8a, 0x2b, 0x17, 0xfe, 0x06,
-	0x33, 0x77, 0x6e, 0xa7, 0x14, 0xa6, 0xa5, 0x2d, 0xb2, 0xeb, 0xcc, 0x3d, 0xe7, 0x3b, 0xdf, 0xf7,
-	0xcd, 0xb9, 0x5f, 0x0a, 0x29, 0x5a, 0x34, 0x74, 0xec, 0x38, 0x16, 0x35, 0xb0, 0x47, 0x99, 0xcd,
-	0x75, 0x17, 0x7b, 0xa4, 0x60, 0xd1, 0x0a, 0xf5, 0xa8, 0x5d, 0xd2, 0xb7, 0x32, 0xfa, 0x66, 0x95,
-	0xb8, 0xb5, 0xb4, 0xe3, 0x32, 0x8f, 0xa1, 0x29, 0x5a, 0x34, 0xd2, 0x47, 0xcb, 0xd3, 0x2d, 0xe5,
-	0xe9, 0xad, 0x8c, 0x9a, 0x28, 0xb1, 0x12, 0x13, 0xd5, 0xba, 0xff, 0x2b, 0x68, 0x54, 0xaf, 0x94,
-	0x18, 0x2b, 0x59, 0x44, 0xc7, 0x0e, 0xd5, 0xb1, 0x6d, 0x33, 0x4f, 0xb6, 0x07, 0xa7, 0x77, 0x4f,
-	0x67, 0xd1, 0x3a, 0x47, 0xb4, 0x69, 0x97, 0x61, 0xfc, 0x85, 0x4f, 0x6e, 0xd1, 0xb2, 0xf2, 0xd8,
-	0x23, 0xab, 0xfe, 0x29, 0xcf, 0x93, 0xcd, 0x2a, 0xe1, 0x9e, 0xb6, 0x09, 0x6a, 0xd4, 0x21, 0x77,
-	0x98, 0xcd, 0x09, 0x7a, 0x09, 0xc3, 0x4d, 0x44, 0x3e, 0xa6, 0x4c, 0xfe, 0x35, 0x33, 0x3c, 0x7b,
-	0x3b, 0x7d, 0xaa, 0xbc, 0x74, 0x88, 0x95, 0xfd, 0x7b, 0xef, 0xdb, 0x44, 0x2c, 0x0f, 0x6e, 0x08,
-	0xae, 0xbd, 0x85, 0x8b, 0x62, 0x64, 0x58, 0x23, 0xb9, 0xa0, 0x04, 0x0c, 0x9a, 0xc4, 0x66, 0x95,
-	0x31, 0x65, 0x52, 0x99, 0x89, 0xe7, 0x83, 0x07, 0xa4, 0x43, 0xc2, 0x28, 0x63, 0xdb, 0x26, 0x56,
-	0x81, 0xb9, 0x05, 0xc3, 0xa2, 0xc4, 0xf6, 0x0a, 0xd4, 0x1c, 0x1b, 0x10, 0x45, 0xff, 0xcb, 0xb3,
-	0xe7, 0x6e, 0x4e, 0x9c, 0xac, 0x98, 0x1a, 0x81, 0xd1, 0xe3, 0xf8, 0x52, 0xce, 0x53, 0x80, 0x26,
-	0x53, 0x31, 0xa5, 0x47, 0x35, 0xf9, 0x78, 0xa8, 0x43, 0x9b, 0x87, 0x89, 0xd6, 0x31, 0x3c, 0x5b,
-	0xcb, 0x95, 0x31, 0xb5, 0x57, 0x96, 0x1a, 0x82, 0xc6, 0xe1, 0x5f, 0xc3, 0x7f, 0xe3, 0xd3, 0x0d,
-	0x34, 0xfd, 0x23, 0x9e, 0x57, 0x4c, 0x6d, 0x1b, 0x26, 0xdb, 0x77, 0x9f, 0xa7, 0xfb, 0x6f, 0xe0,
-	0x66, 0xd4, 0xe0, 0x16, 0x0f, 0x43, 0x01, 0xed, 0xbc, 0x57, 0xda, 0x79, 0xff, 0x41, 0x81, 0x5b,
-	0xdd, 0xc0, 0x9f, 0xa7, 0x42, 0x4d, 0x5a, 0xbb, 0x68, 0x59, 0x59, 0x0b, 0x1b, 0xef, 0x2c, 0xca,
-	0x3d, 0x62, 0x2e, 0xf9, 0xcb, 0x14, 0xae, 0xfd, 0x43, 0x98, 0xea, 0x50, 0x23, 0xd9, 0x8d, 0xc2,
-	0x90, 0x58, 0xc1, 0x80, 0x58, 0x3c, 0x2f, 0x9f, 0xb4, 0x69, 0xb8, 0xd6, 0x68, 0xde, 0x28, 0x53,
-	0x8f, 0x04, 0xcd, 0x8b, 0xa6, 0xe9, 0x12, 0xce, 0x49, 0x38, 0x63, 0x57, 0x81, 0xeb, 0x9d, 0xeb,
-	0xe4, 0x1c, 0x13, 0x46, 0x70, 0xf0, 0xb2, 0xe0, 0x60, 0xea, 0x36, 0x7c, 0x98, 0xeb, 0xc2, 0x87,
-	0x93, 0xb8, 0xeb, 0x98, 0xba, 0xd2, 0x94, 0xff, 0x70, 0xf3, 0x15, 0x9f, 0xfd, 0x09, 0x30, 0x28,
-	0xe8, 0xa0, 0x2f, 0x0a, 0x8c, 0xb4, 0xdc, 0x77, 0x34, 0xdf, 0xc5, 0xa8, 0xb6, 0x19, 0xa2, 0x2e,
-	0xf4, 0xd9, 0x1d, 0xc8, 0xd7, 0x52, 0x1f, 0xbf, 0xfe, 0xf8, 0x3c, 0x70, 0x03, 0x4d, 0xeb, 0x32,
-	0xdf, 0x82, 0x5c, 0x4b, 0x1d, 0xcf, 0xb5, 0x60, 0x49, 0xd0, 0x81, 0x02, 0xf1, 0x10, 0x05, 0xdd,
-	0xef, 0x76, 0xf6, 0xf1, 0xb4, 0x51, 0xe7, 0xfa, 0xe8, 0x94, 0x8c, 0x5f, 0x09, 0xc6, 0xeb, 0x68,
-	0xad, 0x1b, 0xc6, 0x47, 0x7e, 0xbd, 0x8f, 0xba, 0x4c, 0x3b, 0x7a, 0xb1, 0x56, 0x08, 0xa2, 0xae,
-	0xae, 0xc0, 0x85, 0x88, 0x40, 0x40, 0xd9, 0x9e, 0xa9, 0x9e, 0xc8, 0x22, 0x35, 0x77, 0x26, 0x0c,
-	0x29, 0x3c, 0x2b, 0x84, 0xcf, 0xa3, 0x07, 0xbd, 0x09, 0xe7, 0x42, 0xb9, 0xc8, 0xc1, 0x1d, 0xb4,
-	0x3b, 0x00, 0x57, 0x3b, 0xa6, 0x03, 0x5a, 0xed, 0x93, 0x6a, 0x64, 0x86, 0xa9, 0xcf, 0xfe, 0x10,
-	0x9a, 0xb4, 0x60, 0x4d, 0x58, 0xf0, 0x04, 0x2d, 0xf7, 0x63, 0xc1, 0xc9, 0x8f, 0xef, 0x7f, 0xf3,
-	0x44, 0x54, 0x0a, 0xa1, 0x5c, 0x0f, 0xb7, 0xaa, 0x5d, 0xce, 0xa9, 0x4b, 0x67, 0x03, 0x91, 0x9a,
-	0x1f, 0x09, 0xcd, 0x73, 0xe8, 0x5e, 0x57, 0x9a, 0x8b, 0x4d, 0x9c, 0x60, 0xaf, 0x39, 0xfa, 0xa5,
-	0xc0, 0xa5, 0x36, 0x29, 0x88, 0x96, 0x7b, 0xa0, 0xd8, 0x21, 0x6e, 0xd5, 0xc7, 0x67, 0xc6, 0xe9,
-	0x6b, 0xc9, 0xb7, 0x9b, 0x50, 0x05, 0xdc, 0xc0, 0xca, 0x6e, 0xec, 0xd5, 0x93, 0xca, 0x7e, 0x3d,
-	0xa9, 0x7c, 0xaf, 0x27, 0x95, 0x4f, 0x87, 0xc9, 0xd8, 0xfe, 0x61, 0x32, 0x76, 0x70, 0x98, 0x8c,
-	0xbd, 0x5e, 0x28, 0x51, 0xaf, 0x5c, 0x2d, 0xa6, 0x0d, 0x56, 0xd1, 0x0d, 0xc6, 0x2b, 0x8c, 0xfb,
-	0x63, 0x52, 0x25, 0xa6, 0x6f, 0x65, 0x32, 0x7a, 0x85, 0x99, 0x55, 0x8b, 0xf0, 0xa8, 0xa9, 0x5e,
-	0xcd, 0x21, 0xbc, 0x38, 0x24, 0xfe, 0xd3, 0xdd, 0xf9, 0x1d, 0x00, 0x00, 0xff, 0xff, 0xf7, 0x7e,
-	0xc7, 0x77, 0x92, 0x0a, 0x00, 0x00,
+	// 835 bytes of a gzipped FileDescriptorProto
+	0x1f, 0x8b, 0x08, 0x00, 0x00, 0x00, 0x00, 0x00, 0x02, 0xff, 0xd4, 0x57, 0x4f, 0x4f, 0x33, 0x45,
+	0x1c, 0xee, 0x54, 0x41, 0x3b, 0xc8, 0xc1, 0xb1, 0x22, 0x6c, 0xb4, 0x40, 0xfd, 0x03, 0x12, 0xbb,
+	0x93, 0x62, 0x8c, 0x62, 0x20, 0x86, 0x96, 0x80, 0x44, 0x44, 0x5c, 0x13, 0x49, 0x3c, 0xb8, 0x99,
+	0xdd, 0x9d, 0x6c, 0x47, 0xb7, 0x3b, 0xcb, 0xce, 0xb6, 0xa6, 0x31, 0x5c, 0x0c, 0x47, 0x0f, 0x26,
+	0x7e, 0x05, 0x3f, 0x85, 0x5e, 0x3c, 0x12, 0x4f, 0x24, 0xc6, 0x84, 0x93, 0x31, 0xc5, 0xd3, 0x7b,
+	0x78, 0x3f, 0xc3, 0x9b, 0x9d, 0x9d, 0xb6, 0xb4, 0xb4, 0xa5, 0x2d, 0xbd, 0x70, 0xdb, 0x9d, 0xe9,
+	0xef, 0xf9, 0x3d, 0xcf, 0xb3, 0xbf, 0x99, 0x27, 0x85, 0x05, 0x66, 0xd9, 0x98, 0x04, 0x81, 0xc7,
+	0x6c, 0x12, 0x31, 0xee, 0x0b, 0x1c, 0x92, 0x88, 0x9a, 0x1e, 0xab, 0xb2, 0x88, 0xf9, 0x2e, 0xae,
+	0x17, 0xf1, 0x59, 0x8d, 0x86, 0x0d, 0x3d, 0x08, 0x79, 0xc4, 0xd1, 0x2a, 0xb3, 0x6c, 0xfd, 0xf6,
+	0xcf, 0xf5, 0xae, 0x9f, 0xeb, 0xf5, 0xa2, 0xb6, 0x61, 0x73, 0x51, 0xe5, 0x02, 0x5b, 0x44, 0xd0,
+	0xa4, 0x16, 0xd7, 0x8b, 0x16, 0x8d, 0x48, 0x11, 0x07, 0xc4, 0x65, 0xbe, 0xac, 0x4b, 0xe0, 0xb4,
+	0xac, 0xcb, 0x5d, 0x2e, 0x1f, 0x71, 0xfc, 0xa4, 0x56, 0x5f, 0x77, 0x39, 0x77, 0x3d, 0x8a, 0x49,
+	0xc0, 0x30, 0xf1, 0x7d, 0x1e, 0xa9, 0x56, 0xc9, 0xee, 0x07, 0xf7, 0x33, 0xee, 0xe6, 0x24, 0xcb,
+	0xf2, 0x36, 0x5c, 0xfa, 0x32, 0x26, 0xb3, 0xeb, 0x79, 0x06, 0x89, 0xe8, 0x51, 0xbc, 0x2b, 0x0c,
+	0x7a, 0x56, 0xa3, 0x22, 0x42, 0xfb, 0x10, 0x76, 0xb8, 0x2d, 0x82, 0x15, 0xb0, 0x3e, 0xb7, 0xf9,
+	0x8e, 0x9e, 0x08, 0xd1, 0x63, 0x21, 0x7a, 0x62, 0x82, 0x12, 0xa2, 0x9f, 0x10, 0x97, 0xaa, 0x5a,
+	0xe3, 0x56, 0x65, 0xfe, 0x77, 0x00, 0xb5, 0x7e, 0x5d, 0x44, 0xc0, 0x7d, 0x41, 0xd1, 0x57, 0x70,
+	0xae, 0x43, 0x4d, 0x2c, 0x82, 0x95, 0xe7, 0xd6, 0xe7, 0x36, 0xdf, 0xd3, 0xef, 0xf5, 0x54, 0x6f,
+	0x63, 0x95, 0x9e, 0xbf, 0xfc, 0x77, 0x39, 0x65, 0xc0, 0xb0, 0x0d, 0x8e, 0x0e, 0xba, 0xb8, 0xa7,
+	0x25, 0xf7, 0xb5, 0x7b, 0xb9, 0x27, 0x8c, 0xba, 0xc8, 0x7f, 0x0b, 0x5f, 0x95, 0xdc, 0xdb, 0xcd,
+	0x5a, 0xee, 0x64, 0xe1, 0x8c, 0x43, 0x7d, 0x5e, 0x95, 0xc6, 0x64, 0x8c, 0xe4, 0x05, 0x61, 0x98,
+	0xb5, 0x2b, 0xc4, 0xf7, 0xa9, 0x67, 0xf2, 0xd0, 0xb4, 0x3d, 0x46, 0xfd, 0xc8, 0x64, 0x8e, 0x64,
+	0x90, 0x31, 0x5e, 0x56, 0x7b, 0x5f, 0x84, 0x65, 0xb9, 0x73, 0xe8, 0xe4, 0x29, 0x5c, 0xe8, 0xc5,
+	0x57, 0xbe, 0x7c, 0x06, 0x61, 0x47, 0xb2, 0xb2, 0x7f, 0x2c, 0x5b, 0x8c, 0x4c, 0xdb, 0x90, 0xfc,
+	0x05, 0x80, 0xcb, 0xdd, 0x7d, 0x44, 0xa9, 0x51, 0xae, 0x10, 0xe6, 0x1f, 0xee, 0xb5, 0x14, 0x2d,
+	0xc1, 0x17, 0xed, 0x78, 0x25, 0xe6, 0x9b, 0x88, 0x7a, 0x41, 0xbe, 0x1f, 0x3a, 0x3d, 0xa3, 0x90,
+	0x9e, 0x78, 0x14, 0xfe, 0x04, 0x70, 0x65, 0x30, 0x8d, 0x47, 0x31, 0x10, 0xbf, 0x01, 0xf8, 0x6e,
+	0x3f, 0x09, 0x5d, 0xdf, 0xb5, 0xed, 0xe9, 0xa0, 0x79, 0x00, 0x03, 0xe6, 0x61, 0x6a, 0x4e, 0xff,
+	0x05, 0xe0, 0xc6, 0x28, 0x34, 0x1f, 0x85, 0xe7, 0xdf, 0xa9, 0xa9, 0xd9, 0xf5, 0xbc, 0x92, 0x47,
+	0xec, 0xef, 0x3d, 0x26, 0x22, 0xea, 0xec, 0xc5, 0x27, 0x6e, 0xea, 0xb7, 0xd5, 0x05, 0x80, 0xab,
+	0x43, 0x9a, 0x29, 0xbf, 0x16, 0xe0, 0xac, 0x3c, 0xf0, 0x89, 0x55, 0x19, 0x43, 0xbd, 0x4d, 0x4f,
+	0x72, 0x15, 0xbe, 0xd9, 0x62, 0x71, 0x5a, 0x61, 0x11, 0x4d, 0x58, 0xec, 0x3a, 0x4e, 0x48, 0x85,
+	0xa0, 0x53, 0x57, 0xfd, 0x0f, 0x80, 0x6f, 0x0d, 0xef, 0xa7, 0x84, 0x3b, 0x70, 0x9e, 0x24, 0x8b,
+	0x66, 0x40, 0x58, 0xd8, 0x1a, 0x95, 0xad, 0x11, 0x46, 0xe5, 0x2e, 0xee, 0x09, 0x61, 0xa1, 0x9a,
+	0x9b, 0x97, 0x48, 0x67, 0x69, 0x7a, 0x36, 0x6e, 0x3e, 0x81, 0x70, 0x46, 0xea, 0x42, 0x7f, 0x00,
+	0x38, 0xdf, 0x15, 0x40, 0x68, 0x7b, 0x04, 0xce, 0x03, 0xd3, 0x51, 0xdb, 0x99, 0xb0, 0x3a, 0x21,
+	0x99, 0x2f, 0xfc, 0xf4, 0xf7, 0xff, 0xbf, 0xa6, 0xd7, 0xd0, 0xdb, 0x58, 0x25, 0x77, 0x92, 0xd8,
+	0x85, 0xde, 0xc4, 0x4e, 0x0e, 0x24, 0xba, 0x06, 0x30, 0xd3, 0x46, 0x41, 0x1f, 0x8d, 0xda, 0xbb,
+	0x37, 0xb5, 0xb4, 0xad, 0x09, 0x2a, 0x15, 0xe3, 0xaf, 0x25, 0xe3, 0x13, 0x74, 0x3c, 0x0a, 0xe3,
+	0x5b, 0x4f, 0x3f, 0xf6, 0xbb, 0x00, 0xcf, 0xb1, 0xd5, 0x30, 0x93, 0xc8, 0x6c, 0x02, 0xf8, 0x4a,
+	0x9f, 0x38, 0x40, 0xa5, 0xb1, 0xa9, 0xde, 0x89, 0x34, 0xad, 0xfc, 0x20, 0x0c, 0x25, 0xbc, 0x24,
+	0x85, 0x6f, 0xa3, 0x8f, 0xc7, 0x13, 0x2e, 0xa4, 0x72, 0x19, 0xa7, 0xe7, 0xe8, 0xe7, 0x34, 0x7c,
+	0x63, 0xe8, 0x4d, 0x8c, 0x8e, 0x26, 0xa4, 0xda, 0x37, 0x77, 0xb4, 0xcf, 0xa7, 0x84, 0xa6, 0x2c,
+	0x38, 0x96, 0x16, 0x7c, 0x8a, 0xf6, 0x27, 0xb1, 0xe0, 0xee, 0xc7, 0x8f, 0xbf, 0x79, 0xb6, 0xdf,
+	0xfd, 0x8a, 0xca, 0x63, 0x9c, 0xaa, 0x41, 0x51, 0xa0, 0xed, 0x3d, 0x0c, 0x44, 0x69, 0xfe, 0x44,
+	0x6a, 0xde, 0x42, 0x1f, 0x8e, 0xa4, 0xd9, 0xea, 0xe0, 0x98, 0x2a, 0x0b, 0x9e, 0x02, 0xf8, 0xda,
+	0x80, 0xeb, 0x14, 0xed, 0x8f, 0x41, 0x71, 0xc8, 0xfd, 0xaf, 0x1d, 0x3c, 0x18, 0x67, 0xa2, 0x21,
+	0xff, 0xa1, 0x03, 0x65, 0x92, 0x16, 0x56, 0xe9, 0xf4, 0xb2, 0x99, 0x03, 0x57, 0xcd, 0x1c, 0xf8,
+	0xaf, 0x99, 0x03, 0xbf, 0xdc, 0xe4, 0x52, 0x57, 0x37, 0xb9, 0xd4, 0xf5, 0x4d, 0x2e, 0xf5, 0xcd,
+	0x8e, 0xcb, 0xa2, 0x4a, 0xcd, 0xd2, 0x6d, 0x5e, 0xc5, 0xea, 0x9f, 0x10, 0xb3, 0xec, 0x82, 0xcb,
+	0x71, 0xbd, 0x58, 0xc4, 0x55, 0xee, 0xd4, 0x3c, 0x2a, 0xfa, 0x75, 0x8d, 0x1a, 0x01, 0x15, 0xd6,
+	0xac, 0xfc, 0xb7, 0xf2, 0xfe, 0xb3, 0x00, 0x00, 0x00, 0xff, 0xff, 0x42, 0x78, 0xed, 0xeb, 0x98,
+	0x0d, 0x00, 0x00,
 }
 
 // Reference imports to suppress errors if they are not otherwise used.
@@ -647,7 +734,7 @@ type QueryClient interface {
 	AllRateLimits(ctx context.Context, in *QueryAllRateLimitsRequest, opts ...grpc.CallOption) (*QueryAllRateLimitsResponse, error)
 	// Queries a specific rate limit by channel ID and denom
 	// Ex:
-	//  - /ratelimit/{channel_or_client_id}/by_denom?denom={denom}
+	//   - /ratelimit/{channel_or_client_id}/by_denom?denom={denom}
 	RateLimit(ctx context.Context, in *QueryRateLimitRequest, opts ...grpc.CallOption) (*QueryRateLimitResponse, error)
 	// Queries all the rate limits for a given chain
 	RateLimitsByChainID(ctx context.Context, in *QueryRateLimitsByChainIDRequest, opts ...grpc.CallOption) (*QueryRateLimitsByChainIDResponse, error)
@@ -727,7 +814,7 @@ type QueryServer interface {
 	AllRateLimits(context.Context, *QueryAllRateLimitsRequest) (*QueryAllRateLimitsResponse, error)
 	// Queries a specific rate limit by channel ID and denom
 	// Ex:
-	//  - /ratelimit/{channel_or_client_id}/by_denom?denom={denom}
+	//   - /ratelimit/{channel_or_client_id}/by_denom?denom={denom}
 	RateLimit(context.Context, *QueryRateLimitRequest) (*QueryRateLimitResponse, error)
 	// Queries all the rate limits for a given chain
 	RateLimitsByChainID(context.Context, *QueryRateLimitsByChainIDRequest) (*QueryRateLimitsByChainIDResponse, error)
@@ -928,6 +1015,18 @@ func (m *QueryAllRateLimitsRequest) MarshalToSizedBuffer(dAtA []byte) (int, erro
 	_ = i
 	var l int
 	_ = l
+	if m.Pagination != nil {
+		{
+			size, err := m.Pagination.MarshalToSizedBuffer(dAtA[:i])
+			if err != nil {
+				return 0, err
+			}
+			i -= size
+			i = encodeVarintQuery(dAtA, i, uint64(size))
+		}
+		i--
+		dAtA[i] = 0xa
+	}
 	return len(dAtA) - i, nil
 }
 
@@ -951,6 +1050,18 @@ func (m *QueryAllRateLimitsResponse) MarshalToSizedBuffer(dAtA []byte) (int, err
 	_ = i
 	var l int
 	_ = l
+	if m.Pagination != nil {
+		{
+			size, err := m.Pagination.MarshalToSizedBuffer(dAtA[:i])
+			if err != nil {
+				return 0, err
+			}
+			i -= size
+			i = encodeVarintQuery(dAtA, i, uint64(size))
+		}
+		i--
+		dAtA[i] = 0x12
+	}
 	if len(m.RateLimits) > 0 {
 		for iNdEx := len(m.RateLimits) - 1; iNdEx >= 0; iNdEx-- {
 			{
@@ -1060,6 +1171,18 @@ func (m *QueryRateLimitsByChainIDRequest) MarshalToSizedBuffer(dAtA []byte) (int
 	_ = i
 	var l int
 	_ = l
+	if m.Pagination != nil {
+		{
+			size, err := m.Pagination.MarshalToSizedBuffer(dAtA[:i])
+			if err != nil {
+				return 0, err
+			}
+			i -= size
+			i = encodeVarintQuery(dAtA, i, uint64(size))
+		}
+		i--
+		dAtA[i] = 0x12
+	}
 	if len(m.ChainId) > 0 {
 		i -= len(m.ChainId)
 		copy(dAtA[i:], m.ChainId)
@@ -1090,6 +1213,18 @@ func (m *QueryRateLimitsByChainIDResponse) MarshalToSizedBuffer(dAtA []byte) (in
 	_ = i
 	var l int
 	_ = l
+	if m.Pagination != nil {
+		{
+			size, err := m.Pagination.MarshalToSizedBuffer(dAtA[:i])
+			if err != nil {
+				return 0, err
+			}
+			i -= size
+			i = encodeVarintQuery(dAtA, i, uint64(size))
+		}
+		i--
+		dAtA[i] = 0x12
+	}
 	if len(m.RateLimits) > 0 {
 		for iNdEx := len(m.RateLimits) - 1; iNdEx >= 0; iNdEx-- {
 			{
@@ -1127,6 +1262,18 @@ func (m *QueryRateLimitsByChannelOrClientIDRequest) MarshalToSizedBuffer(dAtA []
 	_ = i
 	var l int
 	_ = l
+	if m.Pagination != nil {
+		{
+			size, err := m.Pagination.MarshalToSizedBuffer(dAtA[:i])
+			if err != nil {
+				return 0, err
+			}
+			i -= size
+			i = encodeVarintQuery(dAtA, i, uint64(size))
+		}
+		i--
+		dAtA[i] = 0x12
+	}
 	if len(m.ChannelOrClientId) > 0 {
 		i -= len(m.ChannelOrClientId)
 		copy(dAtA[i:], m.ChannelOrClientId)
@@ -1157,6 +1304,18 @@ func (m *QueryRateLimitsByChannelOrClientIDResponse) MarshalToSizedBuffer(dAtA [
 	_ = i
 	var l int
 	_ = l
+	if m.Pagination != nil {
+		{
+			size, err := m.Pagination.MarshalToSizedBuffer(dAtA[:i])
+			if err != nil {
+				return 0, err
+			}
+			i -= size
+			i = encodeVarintQuery(dAtA, i, uint64(size))
+		}
+		i--
+		dAtA[i] = 0x12
+	}
 	if len(m.RateLimits) > 0 {
 		for iNdEx := len(m.RateLimits) - 1; iNdEx >= 0; iNdEx-- {
 			{
@@ -1194,6 +1353,18 @@ func (m *QueryAllBlacklistedDenomsRequest) MarshalToSizedBuffer(dAtA []byte) (in
 	_ = i
 	var l int
 	_ = l
+	if m.Pagination != nil {
+		{
+			size, err := m.Pagination.MarshalToSizedBuffer(dAtA[:i])
+			if err != nil {
+				return 0, err
+			}
+			i -= size
+			i = encodeVarintQuery(dAtA, i, uint64(size))
+		}
+		i--
+		dAtA[i] = 0xa
+	}
 	return len(dAtA) - i, nil
 }
 
@@ -1217,6 +1388,18 @@ func (m *QueryAllBlacklistedDenomsResponse) MarshalToSizedBuffer(dAtA []byte) (i
 	_ = i
 	var l int
 	_ = l
+	if m.Pagination != nil {
+		{
+			size, err := m.Pagination.MarshalToSizedBuffer(dAtA[:i])
+			if err != nil {
+				return 0, err
+			}
+			i -= size
+			i = encodeVarintQuery(dAtA, i, uint64(size))
+		}
+		i--
+		dAtA[i] = 0x12
+	}
 	if len(m.Denoms) > 0 {
 		for iNdEx := len(m.Denoms) - 1; iNdEx >= 0; iNdEx-- {
 			i -= len(m.Denoms[iNdEx])
@@ -1249,6 +1432,18 @@ func (m *QueryAllWhitelistedAddressesRequest) MarshalToSizedBuffer(dAtA []byte) 
 	_ = i
 	var l int
 	_ = l
+	if m.Pagination != nil {
+		{
+			size, err := m.Pagination.MarshalToSizedBuffer(dAtA[:i])
+			if err != nil {
+				return 0, err
+			}
+			i -= size
+			i = encodeVarintQuery(dAtA, i, uint64(size))
+		}
+		i--
+		dAtA[i] = 0xa
+	}
 	return len(dAtA) - i, nil
 }
 
@@ -1272,6 +1467,18 @@ func (m *QueryAllWhitelistedAddressesResponse) MarshalToSizedBuffer(dAtA []byte)
 	_ = i
 	var l int
 	_ = l
+	if m.Pagination != nil {
+		{
+			size, err := m.Pagination.MarshalToSizedBuffer(dAtA[:i])
+			if err != nil {
+				return 0, err
+			}
+			i -= size
+			i = encodeVarintQuery(dAtA, i, uint64(size))
+		}
+		i--
+		dAtA[i] = 0x12
+	}
 	if len(m.AddressPairs) > 0 {
 		for iNdEx := len(m.AddressPairs) - 1; iNdEx >= 0; iNdEx-- {
 			{
@@ -1306,6 +1513,10 @@ func (m *QueryAllRateLimitsRequest) Size() (n int) {
 	}
 	var l int
 	_ = l
+	if m.Pagination != nil {
+		l = m.Pagination.Size()
+		n += 1 + l + sovQuery(uint64(l))
+	}
 	return n
 }
 
@@ -1320,6 +1531,10 @@ func (m *QueryAllRateLimitsResponse) Size() (n int) {
 			l = e.Size()
 			n += 1 + l + sovQuery(uint64(l))
 		}
+	}
+	if m.Pagination != nil {
+		l = m.Pagination.Size()
+		n += 1 + l + sovQuery(uint64(l))
 	}
 	return n
 }
@@ -1364,6 +1579,10 @@ func (m *QueryRateLimitsByChainIDRequest) Size() (n int) {
 	if l > 0 {
 		n += 1 + l + sovQuery(uint64(l))
 	}
+	if m.Pagination != nil {
+		l = m.Pagination.Size()
+		n += 1 + l + sovQuery(uint64(l))
+	}
 	return n
 }
 
@@ -1379,6 +1598,10 @@ func (m *QueryRateLimitsByChainIDResponse) Size() (n int) {
 			n += 1 + l + sovQuery(uint64(l))
 		}
 	}
+	if m.Pagination != nil {
+		l = m.Pagination.Size()
+		n += 1 + l + sovQuery(uint64(l))
+	}
 	return n
 }
 
@@ -1390,6 +1613,10 @@ func (m *QueryRateLimitsByChannelOrClientIDRequest) Size() (n int) {
 	_ = l
 	l = len(m.ChannelOrClientId)
 	if l > 0 {
+		n += 1 + l + sovQuery(uint64(l))
+	}
+	if m.Pagination != nil {
+		l = m.Pagination.Size()
 		n += 1 + l + sovQuery(uint64(l))
 	}
 	return n
@@ -1407,6 +1634,10 @@ func (m *QueryRateLimitsByChannelOrClientIDResponse) Size() (n int) {
 			n += 1 + l + sovQuery(uint64(l))
 		}
 	}
+	if m.Pagination != nil {
+		l = m.Pagination.Size()
+		n += 1 + l + sovQuery(uint64(l))
+	}
 	return n
 }
 
@@ -1416,6 +1647,10 @@ func (m *QueryAllBlacklistedDenomsRequest) Size() (n int) {
 	}
 	var l int
 	_ = l
+	if m.Pagination != nil {
+		l = m.Pagination.Size()
+		n += 1 + l + sovQuery(uint64(l))
+	}
 	return n
 }
 
@@ -1431,6 +1666,10 @@ func (m *QueryAllBlacklistedDenomsResponse) Size() (n int) {
 			n += 1 + l + sovQuery(uint64(l))
 		}
 	}
+	if m.Pagination != nil {
+		l = m.Pagination.Size()
+		n += 1 + l + sovQuery(uint64(l))
+	}
 	return n
 }
 
@@ -1440,6 +1679,10 @@ func (m *QueryAllWhitelistedAddressesRequest) Size() (n int) {
 	}
 	var l int
 	_ = l
+	if m.Pagination != nil {
+		l = m.Pagination.Size()
+		n += 1 + l + sovQuery(uint64(l))
+	}
 	return n
 }
 
@@ -1454,6 +1697,10 @@ func (m *QueryAllWhitelistedAddressesResponse) Size() (n int) {
 			l = e.Size()
 			n += 1 + l + sovQuery(uint64(l))
 		}
+	}
+	if m.Pagination != nil {
+		l = m.Pagination.Size()
+		n += 1 + l + sovQuery(uint64(l))
 	}
 	return n
 }
@@ -1493,6 +1740,42 @@ func (m *QueryAllRateLimitsRequest) Unmarshal(dAtA []byte) error {
 			return fmt.Errorf("proto: QueryAllRateLimitsRequest: illegal tag %d (wire type %d)", fieldNum, wire)
 		}
 		switch fieldNum {
+		case 1:
+			if wireType != 2 {
+				return fmt.Errorf("proto: wrong wireType = %d for field Pagination", wireType)
+			}
+			var msglen int
+			for shift := uint(0); ; shift += 7 {
+				if shift >= 64 {
+					return ErrIntOverflowQuery
+				}
+				if iNdEx >= l {
+					return io.ErrUnexpectedEOF
+				}
+				b := dAtA[iNdEx]
+				iNdEx++
+				msglen |= int(b&0x7F) << shift
+				if b < 0x80 {
+					break
+				}
+			}
+			if msglen < 0 {
+				return ErrInvalidLengthQuery
+			}
+			postIndex := iNdEx + msglen
+			if postIndex < 0 {
+				return ErrInvalidLengthQuery
+			}
+			if postIndex > l {
+				return io.ErrUnexpectedEOF
+			}
+			if m.Pagination == nil {
+				m.Pagination = &query.PageRequest{}
+			}
+			if err := m.Pagination.Unmarshal(dAtA[iNdEx:postIndex]); err != nil {
+				return err
+			}
+			iNdEx = postIndex
 		default:
 			iNdEx = preIndex
 			skippy, err := skipQuery(dAtA[iNdEx:])
@@ -1574,6 +1857,42 @@ func (m *QueryAllRateLimitsResponse) Unmarshal(dAtA []byte) error {
 			}
 			m.RateLimits = append(m.RateLimits, RateLimit{})
 			if err := m.RateLimits[len(m.RateLimits)-1].Unmarshal(dAtA[iNdEx:postIndex]); err != nil {
+				return err
+			}
+			iNdEx = postIndex
+		case 2:
+			if wireType != 2 {
+				return fmt.Errorf("proto: wrong wireType = %d for field Pagination", wireType)
+			}
+			var msglen int
+			for shift := uint(0); ; shift += 7 {
+				if shift >= 64 {
+					return ErrIntOverflowQuery
+				}
+				if iNdEx >= l {
+					return io.ErrUnexpectedEOF
+				}
+				b := dAtA[iNdEx]
+				iNdEx++
+				msglen |= int(b&0x7F) << shift
+				if b < 0x80 {
+					break
+				}
+			}
+			if msglen < 0 {
+				return ErrInvalidLengthQuery
+			}
+			postIndex := iNdEx + msglen
+			if postIndex < 0 {
+				return ErrInvalidLengthQuery
+			}
+			if postIndex > l {
+				return io.ErrUnexpectedEOF
+			}
+			if m.Pagination == nil {
+				m.Pagination = &query.PageResponse{}
+			}
+			if err := m.Pagination.Unmarshal(dAtA[iNdEx:postIndex]); err != nil {
 				return err
 			}
 			iNdEx = postIndex
@@ -1859,6 +2178,42 @@ func (m *QueryRateLimitsByChainIDRequest) Unmarshal(dAtA []byte) error {
 			}
 			m.ChainId = string(dAtA[iNdEx:postIndex])
 			iNdEx = postIndex
+		case 2:
+			if wireType != 2 {
+				return fmt.Errorf("proto: wrong wireType = %d for field Pagination", wireType)
+			}
+			var msglen int
+			for shift := uint(0); ; shift += 7 {
+				if shift >= 64 {
+					return ErrIntOverflowQuery
+				}
+				if iNdEx >= l {
+					return io.ErrUnexpectedEOF
+				}
+				b := dAtA[iNdEx]
+				iNdEx++
+				msglen |= int(b&0x7F) << shift
+				if b < 0x80 {
+					break
+				}
+			}
+			if msglen < 0 {
+				return ErrInvalidLengthQuery
+			}
+			postIndex := iNdEx + msglen
+			if postIndex < 0 {
+				return ErrInvalidLengthQuery
+			}
+			if postIndex > l {
+				return io.ErrUnexpectedEOF
+			}
+			if m.Pagination == nil {
+				m.Pagination = &query.PageRequest{}
+			}
+			if err := m.Pagination.Unmarshal(dAtA[iNdEx:postIndex]); err != nil {
+				return err
+			}
+			iNdEx = postIndex
 		default:
 			iNdEx = preIndex
 			skippy, err := skipQuery(dAtA[iNdEx:])
@@ -1940,6 +2295,42 @@ func (m *QueryRateLimitsByChainIDResponse) Unmarshal(dAtA []byte) error {
 			}
 			m.RateLimits = append(m.RateLimits, RateLimit{})
 			if err := m.RateLimits[len(m.RateLimits)-1].Unmarshal(dAtA[iNdEx:postIndex]); err != nil {
+				return err
+			}
+			iNdEx = postIndex
+		case 2:
+			if wireType != 2 {
+				return fmt.Errorf("proto: wrong wireType = %d for field Pagination", wireType)
+			}
+			var msglen int
+			for shift := uint(0); ; shift += 7 {
+				if shift >= 64 {
+					return ErrIntOverflowQuery
+				}
+				if iNdEx >= l {
+					return io.ErrUnexpectedEOF
+				}
+				b := dAtA[iNdEx]
+				iNdEx++
+				msglen |= int(b&0x7F) << shift
+				if b < 0x80 {
+					break
+				}
+			}
+			if msglen < 0 {
+				return ErrInvalidLengthQuery
+			}
+			postIndex := iNdEx + msglen
+			if postIndex < 0 {
+				return ErrInvalidLengthQuery
+			}
+			if postIndex > l {
+				return io.ErrUnexpectedEOF
+			}
+			if m.Pagination == nil {
+				m.Pagination = &query.PageResponse{}
+			}
+			if err := m.Pagination.Unmarshal(dAtA[iNdEx:postIndex]); err != nil {
 				return err
 			}
 			iNdEx = postIndex
@@ -2025,6 +2416,42 @@ func (m *QueryRateLimitsByChannelOrClientIDRequest) Unmarshal(dAtA []byte) error
 			}
 			m.ChannelOrClientId = string(dAtA[iNdEx:postIndex])
 			iNdEx = postIndex
+		case 2:
+			if wireType != 2 {
+				return fmt.Errorf("proto: wrong wireType = %d for field Pagination", wireType)
+			}
+			var msglen int
+			for shift := uint(0); ; shift += 7 {
+				if shift >= 64 {
+					return ErrIntOverflowQuery
+				}
+				if iNdEx >= l {
+					return io.ErrUnexpectedEOF
+				}
+				b := dAtA[iNdEx]
+				iNdEx++
+				msglen |= int(b&0x7F) << shift
+				if b < 0x80 {
+					break
+				}
+			}
+			if msglen < 0 {
+				return ErrInvalidLengthQuery
+			}
+			postIndex := iNdEx + msglen
+			if postIndex < 0 {
+				return ErrInvalidLengthQuery
+			}
+			if postIndex > l {
+				return io.ErrUnexpectedEOF
+			}
+			if m.Pagination == nil {
+				m.Pagination = &query.PageRequest{}
+			}
+			if err := m.Pagination.Unmarshal(dAtA[iNdEx:postIndex]); err != nil {
+				return err
+			}
+			iNdEx = postIndex
 		default:
 			iNdEx = preIndex
 			skippy, err := skipQuery(dAtA[iNdEx:])
@@ -2109,6 +2536,42 @@ func (m *QueryRateLimitsByChannelOrClientIDResponse) Unmarshal(dAtA []byte) erro
 				return err
 			}
 			iNdEx = postIndex
+		case 2:
+			if wireType != 2 {
+				return fmt.Errorf("proto: wrong wireType = %d for field Pagination", wireType)
+			}
+			var msglen int
+			for shift := uint(0); ; shift += 7 {
+				if shift >= 64 {
+					return ErrIntOverflowQuery
+				}
+				if iNdEx >= l {
+					return io.ErrUnexpectedEOF
+				}
+				b := dAtA[iNdEx]
+				iNdEx++
+				msglen |= int(b&0x7F) << shift
+				if b < 0x80 {
+					break
+				}
+			}
+			if msglen < 0 {
+				return ErrInvalidLengthQuery
+			}
+			postIndex := iNdEx + msglen
+			if postIndex < 0 {
+				return ErrInvalidLengthQuery
+			}
+			if postIndex > l {
+				return io.ErrUnexpectedEOF
+			}
+			if m.Pagination == nil {
+				m.Pagination = &query.PageResponse{}
+			}
+			if err := m.Pagination.Unmarshal(dAtA[iNdEx:postIndex]); err != nil {
+				return err
+			}
+			iNdEx = postIndex
 		default:
 			iNdEx = preIndex
 			skippy, err := skipQuery(dAtA[iNdEx:])
@@ -2159,6 +2622,42 @@ func (m *QueryAllBlacklistedDenomsRequest) Unmarshal(dAtA []byte) error {
 			return fmt.Errorf("proto: QueryAllBlacklistedDenomsRequest: illegal tag %d (wire type %d)", fieldNum, wire)
 		}
 		switch fieldNum {
+		case 1:
+			if wireType != 2 {
+				return fmt.Errorf("proto: wrong wireType = %d for field Pagination", wireType)
+			}
+			var msglen int
+			for shift := uint(0); ; shift += 7 {
+				if shift >= 64 {
+					return ErrIntOverflowQuery
+				}
+				if iNdEx >= l {
+					return io.ErrUnexpectedEOF
+				}
+				b := dAtA[iNdEx]
+				iNdEx++
+				msglen |= int(b&0x7F) << shift
+				if b < 0x80 {
+					break
+				}
+			}
+			if msglen < 0 {
+				return ErrInvalidLengthQuery
+			}
+			postIndex := iNdEx + msglen
+			if postIndex < 0 {
+				return ErrInvalidLengthQuery
+			}
+			if postIndex > l {
+				return io.ErrUnexpectedEOF
+			}
+			if m.Pagination == nil {
+				m.Pagination = &query.PageRequest{}
+			}
+			if err := m.Pagination.Unmarshal(dAtA[iNdEx:postIndex]); err != nil {
+				return err
+			}
+			iNdEx = postIndex
 		default:
 			iNdEx = preIndex
 			skippy, err := skipQuery(dAtA[iNdEx:])
@@ -2241,6 +2740,42 @@ func (m *QueryAllBlacklistedDenomsResponse) Unmarshal(dAtA []byte) error {
 			}
 			m.Denoms = append(m.Denoms, string(dAtA[iNdEx:postIndex]))
 			iNdEx = postIndex
+		case 2:
+			if wireType != 2 {
+				return fmt.Errorf("proto: wrong wireType = %d for field Pagination", wireType)
+			}
+			var msglen int
+			for shift := uint(0); ; shift += 7 {
+				if shift >= 64 {
+					return ErrIntOverflowQuery
+				}
+				if iNdEx >= l {
+					return io.ErrUnexpectedEOF
+				}
+				b := dAtA[iNdEx]
+				iNdEx++
+				msglen |= int(b&0x7F) << shift
+				if b < 0x80 {
+					break
+				}
+			}
+			if msglen < 0 {
+				return ErrInvalidLengthQuery
+			}
+			postIndex := iNdEx + msglen
+			if postIndex < 0 {
+				return ErrInvalidLengthQuery
+			}
+			if postIndex > l {
+				return io.ErrUnexpectedEOF
+			}
+			if m.Pagination == nil {
+				m.Pagination = &query.PageResponse{}
+			}
+			if err := m.Pagination.Unmarshal(dAtA[iNdEx:postIndex]); err != nil {
+				return err
+			}
+			iNdEx = postIndex
 		default:
 			iNdEx = preIndex
 			skippy, err := skipQuery(dAtA[iNdEx:])
@@ -2291,6 +2826,42 @@ func (m *QueryAllWhitelistedAddressesRequest) Unmarshal(dAtA []byte) error {
 			return fmt.Errorf("proto: QueryAllWhitelistedAddressesRequest: illegal tag %d (wire type %d)", fieldNum, wire)
 		}
 		switch fieldNum {
+		case 1:
+			if wireType != 2 {
+				return fmt.Errorf("proto: wrong wireType = %d for field Pagination", wireType)
+			}
+			var msglen int
+			for shift := uint(0); ; shift += 7 {
+				if shift >= 64 {
+					return ErrIntOverflowQuery
+				}
+				if iNdEx >= l {
+					return io.ErrUnexpectedEOF
+				}
+				b := dAtA[iNdEx]
+				iNdEx++
+				msglen |= int(b&0x7F) << shift
+				if b < 0x80 {
+					break
+				}
+			}
+			if msglen < 0 {
+				return ErrInvalidLengthQuery
+			}
+			postIndex := iNdEx + msglen
+			if postIndex < 0 {
+				return ErrInvalidLengthQuery
+			}
+			if postIndex > l {
+				return io.ErrUnexpectedEOF
+			}
+			if m.Pagination == nil {
+				m.Pagination = &query.PageRequest{}
+			}
+			if err := m.Pagination.Unmarshal(dAtA[iNdEx:postIndex]); err != nil {
+				return err
+			}
+			iNdEx = postIndex
 		default:
 			iNdEx = preIndex
 			skippy, err := skipQuery(dAtA[iNdEx:])
@@ -2372,6 +2943,42 @@ func (m *QueryAllWhitelistedAddressesResponse) Unmarshal(dAtA []byte) error {
 			}
 			m.AddressPairs = append(m.AddressPairs, WhitelistedAddressPair{})
 			if err := m.AddressPairs[len(m.AddressPairs)-1].Unmarshal(dAtA[iNdEx:postIndex]); err != nil {
+				return err
+			}
+			iNdEx = postIndex
+		case 2:
+			if wireType != 2 {
+				return fmt.Errorf("proto: wrong wireType = %d for field Pagination", wireType)
+			}
+			var msglen int
+			for shift := uint(0); ; shift += 7 {
+				if shift >= 64 {
+					return ErrIntOverflowQuery
+				}
+				if iNdEx >= l {
+					return io.ErrUnexpectedEOF
+				}
+				b := dAtA[iNdEx]
+				iNdEx++
+				msglen |= int(b&0x7F) << shift
+				if b < 0x80 {
+					break
+				}
+			}
+			if msglen < 0 {
+				return ErrInvalidLengthQuery
+			}
+			postIndex := iNdEx + msglen
+			if postIndex < 0 {
+				return ErrInvalidLengthQuery
+			}
+			if postIndex > l {
+				return io.ErrUnexpectedEOF
+			}
+			if m.Pagination == nil {
+				m.Pagination = &query.PageResponse{}
+			}
+			if err := m.Pagination.Unmarshal(dAtA[iNdEx:postIndex]); err != nil {
 				return err
 			}
 			iNdEx = postIndex